requirements.txt
Flask
opencv-python
face_recognition_models
flask_cors
getmac
pymongo
cloudinary
requests
numpy
face_recognition
python-dotenv
gunicorn
opencv-python-headless
